Utility to graft old JGit history onto repository
This script can be executed by a developer to download and graft
on the old JGit history, from before we moved the project to the
eclipse.org namespace and the Eclipse Foundation servers.

Executing this script is only necessary if you need to run log or
blame past the migration boundary, and isn't always recommended when
it comes to pushing objects to a remote server.  As mentioned in the
script, it is best to use a specialized repository with this graft,
not your main work repository.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>

Former-commit-id: baba6a3c28718086fb2e6f6d195ff8ae7870b3fa