Fix SUBMITTING_PATCHES to follow the Eclipse IP process
I somehow missed that SUBMITTING_PATCHES referenced the older "send
to git mailing list by email" patch process we used for the first
few years of the JGit project's life.  Now that we are hosted under
the Eclipse Foundation umbrella we need to follow their IP process
for all contributions.

Change-Id: Ia29b31746826ca5b914e0a1d108c2166ff64d268
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>

Former-commit-id: f4964358b3bfdb206ea525042673998690bbfca0