Remove TODO file and move to bugzilla
There is no longer any value in keeping track of things we need to
implement in a file in the top level of the source code.  Time has
proven we don't keep it current, as some of these features are
already implemented, and some are still being worked on.

Worse, many of these items were for the EGit plugin, which has no
relationship to JGit.

Change-Id: Iaa83491819518a1293810aafb566354bdc4b5e28
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>

Former-commit-id: f5871e16d48b5ead71534430e39e7d6ad9958a40