Script to fix license headers and copyrights in Java sources
The script merges explicit copyright statements in all Java
sources with author information from git history, updates the
copyright headers accordingly, and updates the license headers
to EDL.  For recognized copyright formats see the test data in
tools/fix-headers.tst.

To fix headers only in the current working directory:

  ./tools/fix-headers.pl

To fix the headers for all revisions (don't do this if you don't
understand the implications of rewriting history) run:

  ./tools/rewrite-history.sh

Authors are mapped to employer copyright statements through a
hardcoded table in the top of the script.  This is a crude but
simple way to list date ranges under which certain changes need
to be attributed to copyright holders other than the author.

Change-Id: I654d758658cded02d91324c385f336bcc57fd85f
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>

Former-commit-id: 578225870ad6530b57af4584a11a0dd0bad0b7a8